Configurando o mapeamento de filmes

+++ Alura.Filmes/Alura.Filmes.App/Negocio/Filme.cs
@@ -6,7 +6,7 @@
         public string Titulo { get; set; }
         public string Descricao { get; set; }
         public string AnoLancamento { get; set; }
-        public int Duracao { get; set; }
+        public short Duracao { get; set; }
 
     }
 }
